Conflicts:
	IMBApertureParser.m
	IMBMetadataTransformer.m
	IMBNode.m
	IMBOutlineView.m
	NSFileManager+iMedia.m

+++ IMBFireFoxParser.m
@@ -233,7 +233,7 @@
 
 //----------------------------------------------------------------------------------------------------------------------
 
-- (NSImage*) iconForFolderID:(int)aFolderID;
+- (NSImage*) iconForFolderID:(NSInteger)aFolderID;
 {
 	NSImage *result = nil;
 	NSString* filename = nil;
